Refactored TransactionBuilder as Interpreter
